Snake case for files. Move schemas to domain folders
