+++ src/components/slider/index.scss
