+++ src/constants/index.js
@@ -1,6 +1,6 @@
 export * as countries from './countries';
 export * as colors from './colors';
-export * as textMessages from './textMessages';
+export * as textMessages from './text-messages';
 export * as errors from './errors';
 export * as http from './http';
 export * as options from './options';
