+++ src/index.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { Provider } from 'react-redux';
-import App from './App';
+import App from './app';
 import * as serviceWorker from './serviceWorker';
 import store from './store/index';
