+++ src/pages/Login/schema.js
@@ -1,5 +1,5 @@
 import * as Yup from 'yup';
-import { errors } from '../constants';
+import { errors } from '../../constants';
 
 export const LoginValidationSchema = Yup.object().shape({
   email: Yup.string()
