+++ src/pages/Profile/schema.js
@@ -1,5 +1,5 @@
 import * as Yup from 'yup';
-import { errors } from '../constants';
+import { errors } from '../../constants';
 
 export const ProfileValidationSchema = Yup.object()
   .shape({
