--- src/service/utils.js
@@ -1,4 +1,0 @@
-export const customAlert = (message) => {
-  // eslint-disable-next-line no-alert, no-undef
-  alert(message);
-};
