+++ src/store/index.js
@@ -1,5 +1,5 @@
 import { createStore } from 'redux';
-import rootReducer from './reducer/rootReducer';
+import rootReducer from './reducer';
 
 const store = createStore(rootReducer);
