+++ src/store/reducer/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import AuthReducer from './authReducer';
-import HMTReducer from './hmtReducer';
+import AuthReducer from './auth';
+import HMTReducer from './hmt';
 
 const rootReducer = combineReducers({
   auth: AuthReducer,
